+++ backend-rust/CHANGELOG.md
@@ -15,8 +15,8 @@ Database schema version: 12
 ### Fixed
 
 - Fixed bug in indexer where removed delegators still had the restake earnings flag stored as false instead of NULL.
-- Fixed bug in indexer where accumulated pool delegator count is update after delegator is removed (this was possible for protocol version 6 and down).
-- Fixed bug in indexer where CIS-2 transfer events was never recorded, when happening before any other token events (like a Mint), this is considered a bug in the token contract, but no the indexer still record the event.
+- Fixed bug in indexer where accumulated pool delegator count is update after delegator is removed (this was possible for protocol version 6 and earlier).
+- Fixed bug in indexer where CIS-2 transfer events were never recorded when happening before any other token events (like a Mint). This is considered a bug in the token contract, but now the indexer still records these events.
 
 ## [0.1.33] - 2025-03-06